Move components to component folder
