+++ frontend/src/components/common/ReadOnlyUserCard.tsx
